+++ src/milo.js
@@ -12,7 +12,7 @@ var MiloStorage = require('./storage');
 var Datasets  = window.Datasets = require('./datasets');
 var Blockly = window.Blockly = require('milo-blocks');
 var Project = require('./project');
-window.canvas = require('./canvas');
+window.canvas = require('./canvas_');
 
 for (var key in utils) {
   global[key] = utils[key];
