+++ src/plot.js
@@ -197,6 +197,4 @@ Plot.prototype.reactive = function() {
     return gd;
 
 };
-
-
 module.exports = Plot;
